app/requirements.txt
flask
pydantic
python-whois
tranco